# back/docs/swagger.yaml
basePath: /
...
info:
  contact: {}
paths:
  /.well-known/webfinger:
    get:
      parameters:
      - description: search webfinger object by account
        in: query
        name: resource
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              links:
                items:
                  type: object
                type: array
              subject:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: webfinger
      tags:
      - webfinger
...
schemes:
- http
swagger: "2.0"